refactor(project): move branching to separate source file
