+++ crates/project/src/user.rs
@@ -33,41 +33,6 @@ impl Default for UserId {
     }
 }
 
-/// Unique identifier of a branch in a project.
-///
-/// Branches in [`Project`]s are used to indicate (past) branching of a [`Project`].
-/// While leave nodes internally exist as separate [`Project`]s, mearging them combines
-/// the history to a single linear [`Project::log`].
-///
-/// Giving branches a unique [`BranchId`] allows us reconstruct the non linear history from
-/// a single linear log.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[serde(transparent)]
-pub struct BranchId(Uuid);
-
-impl BranchId {
-    /// Creates a new random branch identifier.
-    #[must_use]
-    #[expect(dead_code)]
-    pub fn new() -> Self {
-        Self(Uuid::new_v4())
-    }
-
-    /// Creates a new identifier for the main branch.
-    ///
-    /// The `main` branch has a special constant [`BranchId`].
-    #[must_use]
-    pub const fn main() -> Self {
-        Self(Uuid::from_u128(0))
-    }
-}
-
-impl Default for BranchId {
-    fn default() -> Self {
-        Self::main()
-    }
-}
-
 /// Unique identifier to associate changes with the origin.
 ///
 /// We don't directly associate project changes to a [`UserId`], but to a
